rag_service/requirements.txt
# FastAPI 框架
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# ==================== LangChain 核心 ====================
langchain>=0.1.17
langchain-core>=0.1.43
langchain-community>=0.0.31
langchain-google-genai>=0.0.8
langchain-text-splitters>=0.0.3

# ==================== 向量資料庫 ====================
qdrant-client==1.11.3

# ==================== Embeddings ====================
sentence-transformers==2.3.1

# ==================== 文件載入器 ====================
pypdf==4.0.1
python-docx==1.1.0
python-pptx==0.6.23
openpyxl==3.1.2
unstructured==0.12.0
docx2txt
pdfplumber
tiktoken

# ==================== HTTP 客戶端 ====================
httpx==0.27.0
requests==2.31.0

# ==================== 工具 ====================
python-dotenv==1.0.0
loguru==0.7.2
watchdog==3.0.0

# ==================== 資料處理 ====================
pandas==2.1.4
numpy==1.26.3